fix: module rename to direct url query
