+++ crates/rattler_repodata_gateway/src/gateway/direct_url_query.rs
